+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ CHANGELOG
 Unreleased
 ==========
 
+* Added: Support for Filters API.
+* Added: Support for sort, f, ft, w, and wf parameters on agent list API.
+* Changed: Model AgentList.pagination is now an instance of model FilterPagination.
+* Removed: Model Agent.token.
 
 0.4.0
 =====